refactor(pr07): move settings orchestrator to settings/index.mts (G5f.1 + G5f.4)
core/registration.mts now holds only registerCoreSettings (hidden settings: system migration version). The top-level registerSettings() orchestrator that wires each category's registerXxxSettings() lives in settings/index.mts.

No behavior change. main.mts still imports registerSettings from '@settings'.
