+++ src/settings/core/index.mts
@@ -5,5 +5,5 @@
 export type { FieldChoice, SettingField, SettingsSection } from './_types.mjs';
 export { CORE_KEYS } from './constants.mjs';
 export { registerSettingsMenus } from './menus.mjs';
-export { registerCoreSettings, registerSettings } from './registration.mjs';
+export { registerCoreSettings } from './registration.mjs';
 export { GenericSettingsApp, SettingsStoreSymbol, useSettingsStore } from './sheet/index.mjs';
